# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Orders Service",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/orders": {
      "post": {
        "tags": [
          "order-rest-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderList"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/OrderList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/filter": {
      "get": {
        "tags": [
          "order-rest-controller"
        ],
        "operationId": "filter",
        "parameters": [
          {
            "name": "filterRequest",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FilterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrderDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrderDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationField"
            }
          }
        }
      },
      "ValidationField": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "OrderResponse": {
        "required": [
          "control_number",
          "customer_code",
          "name",
          "unit_price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "control_number": {
            "pattern": "\\d+",
            "type": "string"
          },
          "registration_date": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "unit_price": {
            "type": "number",
            "example": 1
          },
          "quantity": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "customer_code": {
            "pattern": "\\d+",
            "type": "string"
          },
          "total_price": {
            "type": "number"
          }
        }
      },
      "OrderRequest": {
        "required": [
          "control_number",
          "customer_code",
          "name",
          "unit_price"
        ],
        "type": "object",
        "properties": {
          "control_number": {
            "pattern": "\\d+",
            "type": "string"
          },
          "registration_date": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "unit_price": {
            "type": "number"
          },
          "quantity": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "customer_code": {
            "pattern": "\\d+",
            "type": "string"
          }
        }
      },
      "OrderList": {
        "required": [
          "orders"
        ],
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderRequest"
            }
          }
        }
      },
      "FilterRequest": {
        "type": "object",
        "properties": {
          "simplified": {
            "type": "boolean"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "page_size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "string"
          },
          "sort_direction": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "simplified_field": {
            "type": "string"
          },
          "sort_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          "control_number": {
            "pattern": "\\d+",
            "type": "string"
          },
          "start_registration_date": {
            "type": "string",
            "format": "date"
          },
          "end_registration_date": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "unit_price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "customer_code": {
            "pattern": "\\d+",
            "type": "string"
          },
          "total_price": {
            "type": "number"
          }
        }
      },
      "PageOrderDto": {
        "type": "object",
        "properties": {
          "total_pages": {
            "type": "integer",
            "format": "int32"
          },
          "total_elements": {
            "type": "integer",
            "format": "int64"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "number_of_elements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "page_number": {
            "type": "integer",
            "format": "int32"
          },
          "page_size": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "null_handling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignore_case": {
            "type": "boolean"
          }
        }
      }
    }
  }
}